@@ -10,27 +10,31 @@
     "email": "tauraloke@gmail.com",
     "url": "https://github.com/tauraloke"
   },
+  "private": true,
   "scripts": {
-    "postinstall": "electron-builder install-app-deps",
-    "jest": "npx jest .",
-    "lint": "npx eslint .",
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint",
     "test": "npm run lint && npm run jest",
-    "start": "electron .",
-    "dev:clean": "rm -rf ./node_modules/electron/dist/storage/ && rm ./db.sqlite3",
-    "pack": "electron-builder --dir",
-    "dist": "electron-builder --linux --windows",
-    "dist:l": "electron-builder --linux",
-    "dist:w": "electron-builder --windows",
+    "jest": "npx jest .",
+    "jsonize:api": "node ./src/scripts/pack_api.js",
+    "electron:build": "npm run jsonize:api && vue-cli-service electron:build",
+    "electron:serve": "npm run jsonize:api && vue-cli-service electron:serve",
+    "postinstall": "electron-builder install-app-deps",
+    "postuninstall": "electron-builder install-app-deps",
     "release": "np"
   },
+  "main": "background.js",
   "dependencies": {
+    "@electron/remote": "^2.0.8",
+    "core-js": "^3.8.3",
     "electron-context-menu": "^3.5.0",
     "electron-debug": "^3.2.0",
-    "electron-prompt": "^1.7.0",
     "electron-store": "^8.1.0",
     "electron-unhandled": "^4.0.1",
     "electron-updater": "^5.2.1",
     "electron-util": "^0.17.2",
+    "esm": "^3.2.25",
     "exif-parser": "^0.1.12",
     "install": "^0.13.0",
     "iqdb-client": "^2.0.0",
@@ -38,17 +42,30 @@
     "sharp": "^0.30.7",
     "sharp-phash": "^2.0.0",
     "sqlite3": "^5.0.11",
-    "tesseract.js": "^3.0.2"
+    "tesseract.js": "^3.0.2",
+    "vue": "^3.2.13"
   },
   "devDependencies": {
-    "electron": "^20.0.3",
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "electron": "^20.1.0",
     "electron-builder": "^23.3.3",
+    "electron-devtools-installer": "^3.1.0",
     "eslint": "^8.22.0",
     "eslint-plugin-unicorn": "^43.0.2",
-    "jest": "^28.1.3",
+    "eslint-plugin-vue": "^8.0.3",
+    "jest": "^29.0.0",
     "np": "^7.5.0",
     "prettier": "2.7.1",
-    "xo": "^0.51.0"
+    "vue-cli-plugin-electron-builder": "~2.1.1",
+    "xo": "^0.52.1"
+  },
+  "overrides": {
+    "cheerio": "1.0.0-rc.10",
+    "parse5-htmlparser2-tree-adapter": "^6.0.0"
   },
   "xo": {
     "envs": [
@@ -60,40 +77,10 @@
     "publish": false,
     "releaseDraft": false
   },
-  "build": {
-    "appId": "com.electron.Tegownitsa",
-    "asarUnpack": [
-      "libs/sqlite/fuzzy.dll",
-      "**/node_modules/sharp/**"
-    ],
-    "win": {
-      "icon": "build/icon.png"
-    },
-    "mac": {
-      "category": "public.app-category.social-networking",
-      "darkModeSupport": true
-    },
-    "dmg": {
-      "iconSize": 160,
-      "contents": [
-        {
-          "x": 180,
-          "y": 170
-        },
-        {
-          "x": 480,
-          "y": 170,
-          "type": "link",
-          "path": "/Applications"
-        }
-      ]
-    },
-    "linux": {
-      "target": [
-        "AppImage",
-        "deb"
-      ],
-      "category": "Images;Network;Tags"
-    }
-  }
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not dead",
+    "not ie 11"
+  ]
 }
