@@ -1,5 +1,5 @@
 const fs = require('fs');
-const iqdb = require('iqdb-client');
+const iqdb = require('iqdb-client/dist/index.js');
 
 module.exports = {
   run: async (_event, _db, file_path) => {
